# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Closechain AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for the current user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/setup:
    post:
      operationId: setupProject
      tags: [projects]
      summary: Create project with subcontractors and document slots in one step
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupProjectBody"
      responses:
        "201":
          description: Project created with subs and docs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project with details
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{projectId}/approve:
    post:
      operationId: approveProject
      tags: [projects]
      summary: Approve a project closeout package for client portal
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project approved with client portal token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveProjectResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /client-portal/{token}:
    get:
      operationId: getClientPortal
      tags: [client-portal]
      summary: Get client portal data (public, no auth needed)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client portal data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientPortalData"
        "404":
          description: Portal not found or not approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UploadUrlRequest:
      type: object
      required: [name, size, contentType, documentSlotId]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
        documentSlotId:
          type: integer
          description: The document slot this upload is for
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    Project:
      type: object
      required: [id, name, clientName, status, totalDocuments, uploadedDocuments, approvedDocuments, progress, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        jobNumber:
          type: [string, 'null']
        description:
          type: [string, 'null']
        clientName:
          type: string
        address:
          type: [string, 'null']
        endDate:
          type: [string, 'null']
        status:
          type: string
          enum: [active, approved, archived]
        clientPortalToken:
          type: [string, 'null']
        totalDocuments:
          type: integer
        uploadedDocuments:
          type: integer
        approvedDocuments:
          type: integer
        progress:
          type: number
        createdAt:
          type: string
          format: date-time
    ProjectDetail:
      type: object
      required: [id, name, clientName, status, totalDocuments, uploadedDocuments, approvedDocuments, progress, subcontractors, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        jobNumber:
          type: [string, 'null']
        description:
          type: [string, 'null']
        clientName:
          type: string
        address:
          type: [string, 'null']
        endDate:
          type: [string, 'null']
        status:
          type: string
          enum: [active, approved, archived]
        clientPortalToken:
          type: [string, 'null']
        totalDocuments:
          type: integer
        uploadedDocuments:
          type: integer
        approvedDocuments:
          type: integer
        progress:
          type: number
        createdAt:
          type: string
          format: date-time
        subcontractors:
          type: array
          items:
            $ref: "#/components/schemas/SubcontractorWithProgress"
    CreateProjectBody:
      type: object
      required: [name, clientName]
      properties:
        name:
          type: string
          minLength: 1
        jobNumber:
          type: string
        description:
          type: string
        clientName:
          type: string
          minLength: 1
        address:
          type: string
        endDate:
          type: string
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        jobNumber:
          type: string
        description:
          type: string
        clientName:
          type: string
          minLength: 1
        address:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, approved, archived]
...
    SetupProjectBody:
      type: object
      required: [name, clientName, subcontractors]
      properties:
        name:
          type: string
          minLength: 1
        jobNumber:
          type: string
        description:
          type: string
        clientName:
          type: string
          minLength: 1
        address:
          type: string
        endDate:
          type: string
        subcontractors:
          type: array
          items:
            type: object
            required: [vendorName, vendorCode, csiCode, documentTypes]
            properties:
              vendorName:
                type: string
                minLength: 1
              vendorCode:
                type: string
              csiCode:
                type: string
                minLength: 1
              documentTypes:
                type: array
                items:
                  type: string
    ApproveProjectResponse:
      type: object
      required: [clientPortalToken, clientPortalUrl]
      properties:
        clientPortalToken:
          type: string
        clientPortalUrl:
          type: string
    SubcontractorWithProgress:
      type: object
      required: [id, vendorName, vendorCode, csiCode, csiDivision, totalDocuments, uploadedDocuments, approvedDocuments, progress]
      properties:
        id:
          type: integer
        vendorName:
          type: string
        vendorCode:
          type: string
        csiCode:
          type: string
        csiDivision:
          type: string
        totalDocuments:
          type: integer
        uploadedDocuments:
          type: integer
        approvedDocuments:
          type: integer
        progress:
          type: number
...
    ClientPortalData:
      type: object
      required: [projectName, clientName, progress, subcontractors]
      properties:
        projectName:
          type: string
        clientName:
          type: string
        description:
          type: [string, 'null']
        address:
          type: [string, 'null']
        progress:
          type: number
        totalDocuments:
          type: integer
        uploadedDocuments:
          type: integer
        approvedDocuments:
          type: integer
        subcontractors:
          type: array
          items:
            $ref: "#/components/schemas/ClientPortalSubcontractor"
    ClientPortalSubcontractor:
      type: object
      required: [vendorName, csiCode, csiDivision, documents, progress]
      properties:
        vendorName:
          type: string
        csiCode:
          type: string
        csiDivision:
          type: string
        progress:
          type: number
        documents:
          type: array
          items:
            $ref: "#/components/schemas/ClientPortalDocument"
    ClientPortalDocument:
      type: object
      required: [documentType, status]
      properties:
        documentType:
          type: string
        status:
          type: string
          enum: [not_submitted, uploaded, approved]
        fileName:
          type: [string, 'null']
        filePath:
          type: [string, 'null']